renames factorized array module
